+++ src/components/action-popover/action-popover.test.tsx
@@ -31,7 +31,7 @@ afterAll(() => {
 });
 
 afterEach(() => {
-  jest.runAllTimers();
+  jest.runOnlyPendingTimers();
 });
 
 testStyledSystemMarginRTL(
@@ -450,6 +450,7 @@ test("clicking the menu button focuses the first focusable element", async () =>
   );
 
   await user.click(screen.getByRole("button"));
+  jest.runOnlyPendingTimers();
 
   expect(screen.getByRole("button", { name: "example item" })).toBeFocused();
 });
@@ -524,6 +525,7 @@ test.each(["ArrowDown", "Space", "Enter", "ArrowUp"] as const)(
     screen.getByRole("button").focus();
     const userEventKeyCode = key === "Space" ? " " : `{${key}}`;
     await user.keyboard(userEventKeyCode);
+    jest.runOnlyPendingTimers();
 
     expect(screen.getByRole("list")).toBeVisible();
     expect(onOpen).toHaveBeenCalledTimes(1);
@@ -545,6 +547,7 @@ test.each(["ArrowDown", "Space", "Enter"] as const)(
     screen.getByRole("button").focus();
     const userEventKeyCode = key === "Space" ? " " : `{${key}}`;
     await user.keyboard(userEventKeyCode);
+    jest.runOnlyPendingTimers();
 
     expect(
       screen.getByRole("button", { name: "example item 1" })
@@ -552,9 +555,7 @@ test.each(["ArrowDown", "Space", "Enter"] as const)(
   }
 );
 
-// test left in, but skipped, pending investigation/decision on https://github.com/Sage/carbon/issues/6826
-// eslint-disable-next-line jest/no-disabled-tests
-test.skip("pressing ArrowUp key when focused on the menu button selects the last focusable item", async () => {
+test("pressing ArrowUp key when focused on the menu button selects the last focusable item", async () => {
   const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
 
   render(
@@ -566,6 +567,7 @@ test.skip("pressing ArrowUp key when focused on the menu button selects the last
 
   screen.getByRole("button").focus();
   await user.keyboard("{ArrowUp}");
+  jest.runOnlyPendingTimers();
 
   expect(screen.getByRole("button", { name: "example item 2" })).toBeFocused();
 });
@@ -589,12 +591,17 @@ test.each([
 
     screen.getByRole("button").focus();
     await user.keyboard("{Enter}");
+    jest.runOnlyPendingTimers();
+
+    expect(
+      screen.getByRole("button", { name: "example item 1" })
+    ).toBeFocused();
 
     await user.keyboard(keycode);
+    jest.runOnlyPendingTimers();
 
     expect(screen.queryByRole("list")).not.toBeInTheDocument();
     expect(onClose).toHaveBeenCalledTimes(1);
-    // TODO - add test for correct element being focused if functionality is fixed (see FE-6424)
   }
 );
 
@@ -628,15 +635,19 @@ test("pressing the Down Arrow key when the menu is open focuses the next item an
   );
 
   await user.click(screen.getByRole("button"));
+  jest.runOnlyPendingTimers();
   expect(screen.getByRole("button", { name: "example item 1" })).toBeFocused();
 
   await user.keyboard("{ArrowDown}");
+  jest.runOnlyPendingTimers();
   expect(screen.getByRole("button", { name: "example item 2" })).toBeFocused();
 
   await user.keyboard("{ArrowDown}");
+  jest.runOnlyPendingTimers();
   expect(screen.getByRole("button", { name: "example item 3" })).toBeFocused();
 
   await user.keyboard("{ArrowDown}");
+  jest.runOnlyPendingTimers();
   expect(screen.getByRole("button", { name: "example item 1" })).toBeFocused();
 });
 
@@ -652,15 +663,19 @@ test("pressing the Up Arrow key when the menu is open focuses the previous item 
   );
 
   await user.click(screen.getByRole("button"));
+  jest.runOnlyPendingTimers();
   expect(screen.getByRole("button", { name: "example item 1" })).toBeFocused();
 
   await user.keyboard("{ArrowUp}");
+  jest.runOnlyPendingTimers();
   expect(screen.getByRole("button", { name: "example item 3" })).toBeFocused();
 
   await user.keyboard("{ArrowUp}");
+  jest.runOnlyPendingTimers();
   expect(screen.getByRole("button", { name: "example item 2" })).toBeFocused();
 
   await user.keyboard("{ArrowUp}");
+  jest.runOnlyPendingTimers();
   expect(screen.getByRole("button", { name: "example item 1" })).toBeFocused();
 });
 
@@ -677,24 +692,34 @@ test("pressing the Home key when the menu is open focuses the first item, no mat
   );
 
   await user.click(screen.getByRole("button"));
+  jest.runOnlyPendingTimers();
   expect(screen.getByRole("button", { name: "example item 1" })).toBeFocused();
 
   await user.keyboard("{ArrowDown}");
+  jest.runOnlyPendingTimers();
   expect(screen.getByRole("button", { name: "example item 2" })).toBeFocused();
   await user.keyboard("{Home}");
+  jest.runOnlyPendingTimers();
   expect(screen.getByRole("button", { name: "example item 1" })).toBeFocused();
 
   await user.keyboard("{ArrowDown}");
+  jest.runOnlyPendingTimers();
   await user.keyboard("{ArrowDown}");
+  jest.runOnlyPendingTimers();
   expect(screen.getByRole("button", { name: "example item 3" })).toBeFocused();
   await user.keyboard("{Home}");
+  jest.runOnlyPendingTimers();
   expect(screen.getByRole("button", { name: "example item 1" })).toBeFocused();
 
   await user.keyboard("{ArrowDown}");
+  jest.runOnlyPendingTimers();
   await user.keyboard("{ArrowDown}");
+  jest.runOnlyPendingTimers();
   await user.keyboard("{ArrowDown}");
+  jest.runOnlyPendingTimers();
   expect(screen.getByRole("button", { name: "example item 4" })).toBeFocused();
   await user.keyboard("{Home}");
+  jest.runOnlyPendingTimers();
   expect(screen.getByRole("button", { name: "example item 1" })).toBeFocused();
 });
 
@@ -711,20 +736,34 @@ test("pressing the End key when the menu is open focuses the last item, no matte
   );
 
   await user.click(screen.getByRole("button"));
+  jest.runOnlyPendingTimers();
+
   expect(screen.getByRole("button", { name: "example item 1" })).toBeFocused();
 
   await user.keyboard("{End}");
+  jest.runOnlyPendingTimers();
+
   expect(screen.getByRole("button", { name: "example item 4" })).toBeFocused();
 
   await user.keyboard("{ArrowUp}");
+  jest.runOnlyPendingTimers();
+
   expect(screen.getByRole("button", { name: "example item 3" })).toBeFocused();
   await user.keyboard("{End}");
+  jest.runOnlyPendingTimers();
+
   expect(screen.getByRole("button", { name: "example item 4" })).toBeFocused();
 
   await user.keyboard("{ArrowUp}");
+  jest.runOnlyPendingTimers();
+
   await user.keyboard("{ArrowUp}");
+  jest.runOnlyPendingTimers();
+
   expect(screen.getByRole("button", { name: "example item 2" })).toBeFocused();
   await user.keyboard("{End}");
+  jest.runOnlyPendingTimers();
+
   expect(screen.getByRole("button", { name: "example item 4" })).toBeFocused();
 });
 
@@ -740,11 +779,13 @@ test("pressing Space when the menu is open does nothing", async () => {
   );
 
   await user.click(screen.getByRole("button"));
+  jest.runOnlyPendingTimers();
 
   expect(screen.getByRole("list")).toBeVisible();
   expect(screen.getByRole("button", { name: "example item 1" })).toBeFocused();
 
   await user.keyboard(" ");
+  jest.runOnlyPendingTimers();
 
   expect(screen.getByRole("list")).toBeVisible();
   expect(screen.getByRole("button", { name: "example item 1" })).toBeFocused();
@@ -764,21 +805,30 @@ test("pressing an alphabet character when the menu is open selects the next sele
     </ActionPopover>
   );
   await user.click(screen.getByRole("button"));
+  jest.runOnlyPendingTimers();
 
   // moves to first element starting with P
   await user.keyboard("p");
+  jest.runOnlyPendingTimers();
+
   expect(screen.getByRole("button", { name: "Print Invoice" })).toBeFocused();
 
   // moves to next element starting with D - noting that it skips the disabled "Disabled" item
   await user.keyboard("d");
+  jest.runOnlyPendingTimers();
+
   expect(screen.getByRole("button", { name: "Download CSV" })).toBeFocused();
 
   // moves to next element starting with D, it loops to the start
   await user.keyboard("d");
+  jest.runOnlyPendingTimers();
+
   expect(screen.getByRole("button", { name: "Download PDF" })).toBeFocused();
 
   // does nothing when there are no matches
   await user.keyboard("z");
+  jest.runOnlyPendingTimers();
+
   expect(screen.getByRole("button", { name: "Download PDF" })).toBeFocused();
 });
 
@@ -794,11 +844,13 @@ test("pressing a non-printable character key when the menu is open does nothing"
   );
 
   await user.click(screen.getByRole("button"));
+  jest.runOnlyPendingTimers();
 
   expect(screen.getByRole("list")).toBeVisible();
   expect(screen.getByRole("button", { name: "first item" })).toBeFocused();
 
   await user.keyboard("{F1}");
+  jest.runOnlyPendingTimers();
 
   expect(screen.getByRole("list")).toBeVisible();
   expect(screen.getByRole("button", { name: "first item" })).toBeFocused();
@@ -892,7 +944,7 @@ describe("when an item has a submenu with default (left) alignment", () => {
     );
 
     act(() => {
-      jest.runAllTimers();
+      jest.runOnlyPendingTimers();
     });
 
     expect(
@@ -969,7 +1021,7 @@ describe("when an item has a submenu with default (left) alignment", () => {
       screen.getByRole("button", { name: "example item with submenu" })
     );
     act(() => {
-      jest.runAllTimers();
+      jest.runOnlyPendingTimers();
     });
 
     expect(
@@ -978,7 +1030,7 @@ describe("when an item has a submenu with default (left) alignment", () => {
 
     await user.hover(screen.getByRole("button", { name: "example item 2" }));
     act(() => {
-      jest.runAllTimers();
+      jest.runOnlyPendingTimers();
     });
 
     expect(
@@ -1056,9 +1108,11 @@ describe("when an item has a submenu with default (left) alignment", () => {
     );
 
     await user.click(screen.getByRole("button"));
+    jest.runOnlyPendingTimers();
 
     screen.getByRole("button", { name: "example item with submenu" }).focus();
     await user.keyboard("{ArrowLeft}");
+    jest.runOnlyPendingTimers();
 
     const firstItem = screen.getByRole("button", { name: "submenu item 1" });
     expect(firstItem).toBeVisible();
@@ -1127,15 +1181,18 @@ describe("when an item has a submenu with default (left) alignment", () => {
     );
 
     await user.click(screen.getByRole("button"));
+    jest.runOnlyPendingTimers();
 
     screen.getByRole("button", { name: "example item with submenu" }).focus();
     await user.keyboard("{ArrowLeft}");
+    jest.runOnlyPendingTimers();
 
     expect(
       screen.getByRole("button", { name: "submenu item 1" })
     ).toBeVisible();
 
     await user.keyboard("z");
+    jest.runOnlyPendingTimers();
 
     expect(
       screen.getByRole("button", { name: "submenu item 1" })
@@ -1166,9 +1223,11 @@ describe("when an item has a submenu with default (left) alignment", () => {
     );
 
     await user.click(screen.getByRole("button"));
+    jest.runOnlyPendingTimers();
 
     screen.getByRole("button", { name: "example item with submenu" }).focus();
     await user.keyboard("{Enter}");
+    jest.runOnlyPendingTimers();
 
     const firstItem = screen.getByRole("button", { name: "submenu item 1" });
     expect(firstItem).toBeVisible();
@@ -1227,14 +1286,20 @@ describe("when an item has a submenu with default (left) alignment", () => {
     );
 
     await user.click(screen.getByRole("button"));
+    jest.runOnlyPendingTimers();
+
     screen.getByRole("button", { name: "example item with submenu" }).focus();
 
     await user.keyboard("{ArrowLeft}");
+    jest.runOnlyPendingTimers();
+
     expect(
       screen.getByRole("button", { name: "submenu item 1" })
     ).toBeFocused();
 
     await user.keyboard("{Escape}");
+    jest.runOnlyPendingTimers();
+
     expect(screen.queryByRole("list")).not.toBeInTheDocument();
     expect(screen.getByRole("button")).toBeFocused();
   });
@@ -1271,7 +1336,7 @@ describe("when an item has a submenu with default (left) alignment", () => {
     );
 
     act(() => {
-      jest.runAllTimers();
+      jest.runOnlyPendingTimers();
     });
 
     expect(
@@ -1449,9 +1514,11 @@ describe("when there isn't enough space on the screen to render a submenu on the
     );
 
     await user.click(screen.getByRole("button"));
+    jest.runOnlyPendingTimers();
 
     screen.getByRole("button", { name: "example item with submenu" }).focus();
     await user.keyboard("{ArrowRight}");
+    jest.runOnlyPendingTimers();
 
     const firstItem = screen.getByRole("button", { name: "submenu item 1" });
     expect(firstItem).toBeVisible();
@@ -1556,9 +1623,11 @@ describe("when the submenuPosition prop is set to 'right'", () => {
     );
 
     await user.click(screen.getByRole("button"));
+    jest.runOnlyPendingTimers();
 
     screen.getByRole("button", { name: "example item with submenu" }).focus();
     await user.keyboard("{ArrowRight}");
+    jest.runOnlyPendingTimers();
 
     const firstItem = screen.getByRole("button", { name: "submenu item 1" });
     expect(firstItem).toBeVisible();
@@ -1680,9 +1749,11 @@ describe("when the submenuPosition prop is set to 'right' and there isn't enough
     );
 
     await user.click(screen.getByRole("button"));
+    jest.runOnlyPendingTimers();
 
     screen.getByRole("button", { name: "example item with submenu" }).focus();
     await user.keyboard("{ArrowLeft}");
+    jest.runOnlyPendingTimers();
 
     const firstItem = screen.getByRole("button", { name: "submenu item 1" });
     expect(firstItem).toBeVisible();
@@ -1858,11 +1929,13 @@ describe("When ActionPopoverMenu contains multiple disabled items", () => {
     );
 
     await user.click(screen.getByRole("button"));
+    jest.runOnlyPendingTimers();
 
     expect(
       screen.getByRole("button", { name: "example item 1" })
     ).toBeFocused();
     await user.keyboard("{ArrowDown}");
+    jest.runOnlyPendingTimers();
 
     expect(
       screen.getByRole("button", { name: "example item 4" })
@@ -1884,15 +1957,19 @@ describe("When ActionPopoverMenu contains multiple disabled items", () => {
     );
 
     await user.click(screen.getByRole("button"));
+    jest.runOnlyPendingTimers();
 
     expect(
       screen.getByRole("button", { name: "example item 1" })
     ).toBeFocused();
     await user.keyboard("{ArrowDown}");
+    jest.runOnlyPendingTimers();
+
     expect(
       screen.getByRole("button", { name: "example item 4" })
     ).toBeFocused();
     await user.keyboard("{ArrowUp}");
+    jest.runOnlyPendingTimers();
 
     expect(
       screen.getByRole("button", { name: "example item 1" })
@@ -1914,19 +1991,25 @@ describe("When ActionPopoverMenu contains multiple disabled items", () => {
     );
 
     await user.click(screen.getByRole("button"));
+    jest.runOnlyPendingTimers();
 
     expect(
       screen.getByRole("button", { name: "example item 2" })
     ).toBeFocused();
     await user.keyboard("{ArrowDown}");
+    jest.runOnlyPendingTimers();
+
     expect(
       screen.getByRole("button", { name: "example item 4" })
     ).toBeFocused();
     await user.keyboard("{ArrowDown}");
+    jest.runOnlyPendingTimers();
+
     expect(
       screen.getByRole("button", { name: "example item 5" })
     ).toBeFocused();
     await user.keyboard("{Home}");
+    jest.runOnlyPendingTimers();
 
     expect(
       screen.getByRole("button", { name: "example item 2" })
@@ -1948,11 +2031,13 @@ describe("When ActionPopoverMenu contains multiple disabled items", () => {
     );
 
     await user.click(screen.getByRole("button"));
+    jest.runOnlyPendingTimers();
 
     expect(
       screen.getByRole("button", { name: "example item 2" })
     ).toBeFocused();
     await user.keyboard("{End}");
+    jest.runOnlyPendingTimers();
 
     expect(
       screen.getByRole("button", { name: "example item 5" })